+++ Sources/Main/Data/ModelSerializer.cs
@@ -33,7 +33,7 @@ namespace DevOnBike.Overfit.Data
             LoadMatrix(br, fc.Biases.Data);
         }
 
-        private static void SaveMatrix(BinaryWriter bw, FastMatrix<double> matrix)
+        private static void SaveMatrix(BinaryWriter bw, FastMatrix<float> matrix)
         {
             bw.Write(matrix.Rows);
             bw.Write(matrix.Cols);
@@ -44,7 +44,7 @@ namespace DevOnBike.Overfit.Data
             }
         }
 
-        private static void LoadMatrix(BinaryReader br, FastMatrix<double> matrix)
+        private static void LoadMatrix(BinaryReader br, FastMatrix<float> matrix)
         {
             var rows = br.ReadInt32();
             var cols = br.ReadInt32();
@@ -55,9 +55,10 @@ namespace DevOnBike.Overfit.Data
             }
 
             var data = matrix.AsSpan();
+
             for (var i = 0; i < data.Length; i++)
             {
-                data[i] = br.ReadDouble();
+                data[i] = br.ReadSingle();
             }
         }
     }
