+++ Sources/Main/DeepLearning/BatchNorm1D.cs
@@ -7,27 +7,27 @@ namespace DevOnBike.Overfit.DeepLearning
         public AutogradNode Gamma { get; private set; }
         public AutogradNode Beta { get; private set; }
 
-        public FastMatrix<double> RunningMean { get; private set; }
-        public FastMatrix<double> RunningVar { get; private set; }
+        public FastMatrix<float> RunningMean { get; private set; }
+        public FastMatrix<float> RunningVar { get; private set; }
 
-        public double Momentum { get; set; }
-        public double Eps { get; set; }
+        public float Momentum { get; set; }
+        public float Eps { get; set; }
 
         public bool IsTraining { get; private set; } = true;
 
-        public BatchNorm1D(int features, double momentum = 0.1, double eps = 1e-5)
+        public BatchNorm1D(int features, float momentum = 0.1f, float eps = 1e-5f)
         {
-            var gammaData = new FastMatrix<double>(1, features);
-            gammaData.AsSpan().Fill(1.0);
+            var gammaData = new FastMatrix<float>(1, features);
+            gammaData.AsSpan().Fill(1f);
             Gamma = new AutogradNode(gammaData, requiresGrad: true);
 
-            var betaData = new FastMatrix<double>(1, features);
-            betaData.AsSpan().Fill(0.0);
+            var betaData = new FastMatrix<float>(1, features);
+            betaData.AsSpan().Fill(0f);
             Beta = new AutogradNode(betaData, requiresGrad: true);
 
-            RunningMean = new FastMatrix<double>(1, features);
-            RunningVar = new FastMatrix<double>(1, features);
-            RunningVar.AsSpan().Fill(1.0);
+            RunningMean = new FastMatrix<float>(1, features);
+            RunningVar = new FastMatrix<float>(1, features);
+            RunningVar.AsSpan().Fill(1f);
 
             Momentum = momentum;
             Eps = eps;
@@ -37,6 +37,7 @@ namespace DevOnBike.Overfit.DeepLearning
         {
             IsTraining = true;
         }
+
         public void Eval()
         {
             IsTraining = false;
@@ -57,7 +58,6 @@ namespace DevOnBike.Overfit.DeepLearning
         {
             using var fs = new FileStream(path, FileMode.Create);
             using var bw = new BinaryWriter(fs);
-
             Save(bw);
         }
 
@@ -72,26 +72,24 @@ namespace DevOnBike.Overfit.DeepLearning
         public void Load(string path)
         {
             if (!File.Exists(path)) throw new FileNotFoundException($"Brak pliku wag: {path}");
-
             using var fs = new FileStream(path, FileMode.Open);
             using var br = new BinaryReader(fs);
-
             Load(br);
         }
 
         public void Load(BinaryReader br)
         {
             var gSpan = Gamma.Data.AsSpan();
-            for (var i = 0; i < gSpan.Length; i++) gSpan[i] = br.ReadDouble();
+            for (var i = 0; i < gSpan.Length; i++) gSpan[i] = br.ReadSingle();
 
             var bSpan = Beta.Data.AsSpan();
-            for (var i = 0; i < bSpan.Length; i++) bSpan[i] = br.ReadDouble();
+            for (var i = 0; i < bSpan.Length; i++) bSpan[i] = br.ReadSingle();
 
             var rmSpan = RunningMean.AsSpan();
-            for (var i = 0; i < rmSpan.Length; i++) rmSpan[i] = br.ReadDouble();
+            for (var i = 0; i < rmSpan.Length; i++) rmSpan[i] = br.ReadSingle();
 
             var rvSpan = RunningVar.AsSpan();
-            for (var i = 0; i < rvSpan.Length; i++) rvSpan[i] = br.ReadDouble();
+            for (var i = 0; i < rvSpan.Length; i++) rvSpan[i] = br.ReadSingle();
         }
 
         public void Dispose()
