+++ Sources/Main/Optimizers/IOptimizer.cs
@@ -2,11 +2,11 @@ namespace DevOnBike.Overfit.Optimizers
 {
     public interface IOptimizer
     {
-        double LearningRate { get; set; }
-    
+        float LearningRate { get; set; }
+
         // Krok optymalizacji (aktualizacja wag)
         void Step();
-    
+
         // Czyszczenie gradientów przed nowym batchem
         void ZeroGrad();
     }
